Major Refactoring To Implement React With Redux
Application states are now managed in redux store whereby actions are
dispatched to carry out payload and reducers changes the state in
response. States are made sure to never be mutated but instead be made a
copy of when they are being updated.

Additional packages have been installed to achieve this, mainly:
- react-redux (for linking react front-end with redux store)
- redux
- redux-logger (for development)
- redux-thunk (for handling asynch actions)
- babel-plugin-transform-object-rest-spread (for object spread syntax
for states)

Connection to server via the new implementation has not yet been
established in this commit however.